fix: add decorators to Header and InputField story
