+++ src/assets/styles/resetStyles.ts
@@ -49,6 +49,18 @@ const resetStyles = css`
     text-decoration: none;
     color: #333;
   }
+
+  .sr-only {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip-path: inset(50%);
+    border: 0;
+    clip: rect(0 0 0 0);
+  }
 `;
 
 export default resetStyles;
