--- src/components/common/Header/index.stories.ts
@@ -1,14 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import Header from '.';
-
-const meta: Meta<typeof Header> = {
-  title: 'common/Header',
-  component: Header,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof Header>;
-
-export const Default: Story = {};
